[backend] merge : B/wine -> backend

See merge request s04-bigdata-sub3/s04p23c102!3

+++ backend/yourwine/src/main/java/com/ssafy/yourwine/model/dto/WineDTO.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 @Data
 public class WineDTO {
 
-	private int wineId;
+	private Long wineId;
 	private String img;
 	private String kname;
 	private String ename;
