+++ backend/yourwine/src/main/java/com/ssafy/yourwine/model/key/ScrapKey.java
@@ -17,6 +17,6 @@ public class ScrapKey implements Serializable{
 	private String userId;
 	
 	@Column(name = "wine_id")
-	private int wineId;
+	private Long wineId;
 }
